moved scripts and blank theme to lib/ folder and no-name error catch
