+++ lib/blank/partials/basics.handlebars
@@ -1,0 +1,11 @@
+<h2>Contact Information</h2>
+<dl>
+  <dt>E-mail</dt>
+    <dd>{{basics.email}}</dd>
+  <dt>Phone</dt>
+    <dd>{{basics.phone}}</dd>
+  <dt>Address</dt>
+    <dd>{{basics.location.address}}</dd>
+  <dt>Postal Code</dt>
+    <dd>{{basics.location.postalCode}}</dd>
+</dl>
