+++ lib/blank/partials/work.handlebars
@@ -1,0 +1,6 @@
+<h2>Work</h2>
+<ul>
+{{#each work}}
+  <li>{{name}}</li>
+{{/each}}
+</ul>
