+++ lib/blank/resume.handlebars
@@ -1,0 +1,23 @@
+<!doctype html>
+<html lang="nl">
+<head>
+<meta charset="utf-8">
+<meta http-equiv="x-ua-compatible" content="ie=edge">
+
+<title>{{basics.name}} - Curriculum Vitae</title>
+
+<meta name="description" content="{{basics.summary}}">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+
+<link rel="stylesheet" href="css/main.css">
+</head>
+<body>
+
+<h1>{{basics.name}}</h1>
+
+{{> basics}}
+
+{{> work}}
+
+</body>
+</html>
